Fix -Wsign-compare warnings.

+++ sys/arch/news68k/dev/kb_hb.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: kb_hb.c,v 1.3 2002/10/02 04:40:08 thorpej Exp $	*/
+/*	$NetBSD: kb_hb.c,v 1.4 2003/01/11 16:00:48 tsutsui Exp $	*/
 
 /*
  * Copyright (c) 2001 Izumi Tsutsui.
@@ -74,7 +74,7 @@ kb_hb_match(parent, cf, aux)
 		return 0;
 
 	/* XXX no default address */
-	if (ha->ha_address == -1)
+	if (ha->ha_address == (u_int)-1)
 		return 0;
 
 	addr = IIOV(ha->ha_address); /* XXX */
